# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UsersController_create
      parameters: []
      requestBody:
        required: true
        description: New user payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john.doe@example.com
                password:
                  type: string
                  example: Secret123!
                roleId:
                  type: number
                  example: 2
              required:
                - name
                - email
                - password
                - roleId
              additionalProperties: false
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: John Doe
                  email:
                    type: string
                    example: john.doe@example.com
                  roleId:
                    type: number
                    example: 2
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
      security:
        - bearer: []
      tags:
        - Users
    get:
      operationId: UsersController_findAll
      parameters: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: John Doe
                    email:
                      type: string
                      example: john.doe@example.com
                    roleId:
                      type: number
                      example: 2
                    createdAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:00:00.000Z'
                    updatedAt:
                      type: string
                      format: date-time
                      example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Users
  /users/{id}:
    get:
      operationId: UsersController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: User by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: John Doe
                  email:
                    type: string
                    example: john.doe@example.com
                  roleId:
                    type: number
                    example: 2
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:34:56.000Z'
      security:
        - bearer: []
      tags:
        - Users
    patch:
      operationId: UsersController_update
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            example: 1
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Johnny Doe
                email:
                  type: string
                  format: email
                  example: johnny.doe@example.com
                password:
                  type: string
                  example: NewSecret123!
                roleId:
                  type: number
                  example: 1
              additionalProperties: false
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  name:
                    type: string
                    example: Johnny Doe
                  email:
                    type: string
                    example: johnny.doe@example.com
                  roleId:
                    type: number
                    example: 1
                  createdAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:00:00.000Z'
                  updatedAt:
                    type: string
                    format: date-time
                    example: '2025-10-30T12:45:00.000Z'
      security:
        - bearer: []
      tags:
        - Users
    delete:
      operationId: UsersController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: User ID
          schema:
            example: 1
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 1
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Users
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http